fixed ota not connecting to wifi before trying to update

+++ src/app/ota/custom_update.py
@@ -2,6 +2,7 @@ import json
 import machine
 import os
 
+from app import wifi
 from app.ota.ota_updater import OTAUpdater
 
 _TAG_FILE = 'version_queue.txt'
@@ -19,20 +20,23 @@ def handle_update():
 
 def _update_if_queued():
     if _TAG_FILE in os.listdir():
-        with open(_TAG_FILE) as file:
-            version = file.read()
-        os.remove(_TAG_FILE)
-        with open('temp-token.txt') as file:
-            token = json.loads(file.read())
+        # connect to wifi
+        wifi.try_connect()
+        if wifi.is_connected():
+            with open(_TAG_FILE) as file:
+                version = file.read()
+            os.remove(_TAG_FILE)
+            with open('temp-token.txt') as file:
+                token = json.loads(file.read())
 
-        if token is None:
-            ota = OTAUpdater('micro-nova/WallPanel', main_dir='app', github_src_dir='src', module='')
-            print('OTAUpdater loaded without token.')
-        else:
-            ota = OTAUpdater('micro-nova/WallPanel', main_dir='app', github_src_dir='src', module='',
-                              headers={'Authorization': 'token {}'.format(token['token'])})
-            print('OTAUpdater loaded with token.')
+            if token is None:
+                ota = OTAUpdater('micro-nova/WallPanel', main_dir='app', github_src_dir='src', module='')
+                print('OTAUpdater loaded without token.')
+            else:
+                ota = OTAUpdater('micro-nova/WallPanel', main_dir='app', github_src_dir='src', module='',
+                                  headers={'Authorization': 'token {}'.format(token['token'])})
+                print('OTAUpdater loaded with token.')
 
-        ota.install_tagged_release(version)
-        print('resetting machine...')
-        machine.reset()
+            ota.install_tagged_release(version)
+            print('resetting machine...')
+            machine.reset()
